PathPlanner Debug
