+++ src/main/java/frc/robot/RobotContainer.java
@@ -43,7 +43,6 @@ public class RobotContainer {
 
     /* Subsystems */
     private final Swerve s_Swerve = new Swerve();
-    private final PathPlannerExampleSub pathPlannerExampleSub = new PathPlannerExampleSub();
 
 
     /** The container for the robot. Contains subsystems, OI devices, and commands. */
@@ -63,32 +62,6 @@ public class RobotContainer {
 
         //NamedCommands.registerCommand("Example Command", new PathPlannerExampleCommand(pathPlannerExampleSub));
 
-        AutoBuilder.configureHolonomic(
-            s_Swerve::getPose, 
-            s_Swerve::setPose,
-            s_Swerve::getChassisSpeed, 
-            s_Swerve::setChassisSpeed,
-            new HolonomicPathFollowerConfig(
-                new PIDConstants(Constants.Swerve.driveKP, 0.0, 0.0), // Translation PID constants
-                new PIDConstants(5.0, 0.0, 0.0), // Rotation PID constants
-                4.5, // Max module speed, in m/s
-                0.4, // Drive base radius in meters. Distance from robot center to furthest module.
-                new ReplanningConfig() // Default path replanning config. See the API for the options here
-            ),
-                
-            () -> {
-                // Boolean supplier that controls when the path will be mirrored for the red alliance
-                // This will flip the path being followed to the red side of the field.
-                // THE ORIGIN WILL REMAIN ON THE BLUE SIDE
-
-                var alliance = DriverStation.getAlliance();
-                if (alliance.isPresent()) {
-                    return alliance.get() == DriverStation.Alliance.Red;
-                }
-                return false;
-            },
-            s_Swerve
-        );
     }
 
     /**
@@ -112,8 +85,8 @@ public class RobotContainer {
     public Command getAutonomousCommand() {
 
         // An ExampleCommand will run in autonomous
-        PathPlannerPath path = PathPlannerPath.fromPathFile("Example Path123");
+        //PathPlannerPath path = PathPlannerPath.fromPathFile("Example Path123");
 
-        return AutoBuilder.followPath(path);
+        return s_Swerve.followPathCommand("Example Path123");
     }
 }
